Fix #414 - `sbt-devoops-release-version-policy` cannot handle aggregate projects with different cross Scala versions
`sbt-release`, which is used by `sbt-devoops-release-version-policy`, can't handle aggregate projects with different cross Scala versions. ([Issue reproted](https://github.com/sbt/sbt-release/issues/214))

e.g.) So `sbt-release` can't handle a project structured like the following.
- root
  - module1 - Scala 2 and 3
  - module2 - Scala 2

It releases artifacts for only Scala 2.

This PR is a workaround to handle it based on https://github.com/sbt/sbt-release/issues/214#issuecomment-368407088 with a way to skip test if it's set to be skipped.